build(vrt): fix logic for generating snapshot branches from forks (#2957)
* build(vrt): fix branch names from forks

* refactor: adjust naming a bit

* fix: use original branch name for pr

* refactor(vrt): use unique snapshot branch names

We used to append `--vrt` to the original branch name to generate a
snapshot branch. This can lead to a collision in branch names on
`uber/baseweb`, so now we scope the branch name with the owner and repo
name, within which the branch name must be unique.

The `--vrt` bit is also moved to the front of the branch name as a
grouping token to make filtering and searching these automated snapshots
branches a bit easier.