feat(timepicker): Swappable date libraries (#2968)
* feat: date-provider

* fix: correctly initial date utils provider

* fix: convert dateToSeconds

* refactor: rework secondsToHourMinute to use utils

* refactor: rework componentDidMount to use utils

* refactor: rework handleChange to use utils

* refactor: cleanup overwrapped function call

* refactor: swap date instance checks with isValid

The problem here is that with the swapping of date libraries,
it's not longer possible to assume the value will be a pure date
object.  It could be a javascript date object, a moment object,
or whatever.  Instead, date.io provides isValid for the same purpose
purpose

* fix: added flow type checking for date swapping

* fix: fix yarn lock

* refactor: remove instance creation

There's no real advantage to creating the instance
ourselves, and this has the added benenfit of no
longer needing to type the adapter

* fix: set utils object to key in context

* fix: fix issues introduced by type checking

* fix: convert back to native date before calling onChange

This is a temporary change to simplify the type checking in other
areas until we can update other components to expect Date.io
dates as well

* fix: add TS types to DateUtilsProvider

* test: add new scenario / e2e test for using moment

* fix: pass date-io dates into onChange

Letting the user pass in a moment date, and then having a JSDate be
passed to onChange was just a bad idea that was introducing bugs

Returning the generic date-io object meant that any library code that
was assuming the returned object was going to be a JSDate needed
to be made generic as well.

The only instance I could find was these calls in the data-table.
While the ideal solution would be to pass utils into `applyTimeToDate`
, I'm going to hold off on doing that until all of the date functions in
the date picker have been moved over to use utils as well.

Instead I just converted back into a JSDate(which for standard users
won't do anything).

* refactor: make destructuring easier to follow

Co-Authored-By: Chase Starr <chasestarr@gmail.com>

* refactor: swap date handler with props

This pulls out of all of the context depdenty date swapping
functionality and uses props instead

This also adds a dedicated DateHelpers class.  When writing
functions that are dependent on the date adapter, it needs
to be passed all over the place, and correctly typed in each
instance.  For Timepicker, this was annoying but not a huge
deal, but for Datepicker, that would be unworkable.
DateHelpers makes the process of writing adapter dependent
functions much easier.

* Update src/timepicker/timepicker.js

fix: add type to format

Co-Authored-By: Chase Starr <chasestarr@gmail.com>

* fix: ensure that null isn't allowed as a date input

* fix: remove unused format types

* fix: remove old index from index file

* fix: use moment adapter to generate date in test

* fix: use old processing in standard timepicker

* fix: correctly handle format type

* fix: reorder adapter isValue check

this was throwing a type error

* fix: recreate date helpers if adapter changes

* fix: correctly initialize adapter

This was missing new

* fix: update leftover utils calls

* fix: correclty intitialize Moment utils

* test(vrt): update visual snapshots for 8aa46a4 [skip ci]

* fix: make type names consistent

* fix: recreate dateHelpers before building steps

Because buidling steps uses dateHelpers, this could
cause the date handling behavior to get out of sync

* fix: make prop destructuring patterns consistent

Co-authored-by: Chase Starr <chasestarr@gmail.com>
Co-authored-by: UberOpenSourceBot <UberOpenSourceBot@users.noreply.github.com>